- Moved folder - Fixed some issues that turned up during debug.
